feat: Extract event generator from XmlSerializer
